Radio buttons, and check boxes
Created radio button and check box classes that will allow radio buttons
and check boxes to be created throughout the program, particularly
useful for options menu in which there are examples of each for testing
purposes. Feel free to remove or edit the temporary ones, they are just
there to show how they work.