Added more to chat box, commented out some code
Chat box now can have multiple lines that will be printed to the chat
window, but chat window does not yet have a scroll bar, so the chat
window has a fixed number of lines (will crash if goes over). Also
commented out the turn code because it interferes with other normal
interactions of the code currently.