Created square tokens with very basic movement
Created a square token for each character currently playing. Currently
each time a player piece moves, it moves to that location, and if
another player piece is there, it covers it up. There is also a wheel on
each property that has a player. This wheel has each players color who
is currently on that location. It currently has artifacts making it less
than ideal but it is a problem with pygame's rendering of circles and
arcs, so I can't do much about it at the moment. More precise movement
and capabilities will hopefully be added later.