Got movement working
When dice is rolled, token appears where it is positioned, then it will
move through every location, and wait for a second at the last location,
then disappear and leave a popup message for that location. The wheel is
only displayed for inactive characters, and therefore is not displayed
while a token is moving.