Implemented bankruptcy and game ending.
When a player goes bankrupt, their buildings become unowned and their section in the display is grayed out.
When the game ends, an appropriate message is displayed. Upon clicking OK, it returns to the Start Menu. Another game can then be started, but currently it doesn't function correctly if you try to change the players.